refactor settings models namespace
