--- app/models/spid_rails/settings/metadata.rb
@@ -1,11 +1,0 @@
-module SpidRails
-
-  class Settings::Metadata < Settings
-
-    def to_hash
-      sp_attributes
-    end
-
-  end
-  
-end
