--- app/models/spid_rails/settings/slo.rb
@@ -1,13 +1,0 @@
-module SpidRails
-
-  class Settings::Slo < Settings
-
-    def to_hash
-      sso_attributes = sp_attributes.merge(idp_attributes)
-      sso_attributes[:sessionindex] = @session_index
-      sso_attributes
-    end
-
-  end
-
-end
